fix(compile): Making influxd compilable on Windows
